# RFW_OPENAPI.yaml
openapi: 3.0.3
info:
  title: RFW 防火墙 API
  version: 2.0.0
  description: |
    # RFW - 基于 eBPF/XDP 的高性能防火墙 API 文档

    基于 Rust 和 aya 开发的高性能防火墙控制接口。
    支持实时规则同步、GeoIP 自动化下载和 CIDR 匹配。

  contact:
    name: RFW 项目
    url: https://github.com/danbai/rfw
  license:
    name: MIT

servers:
  - url: http://localhost:8080
    description: 本地开发环境

tags:
  - name: 规则管理
    description: 防火墙规则的增删改查
  - name: 系统状态
    description: 获取防火墙运行信息

paths:
  /api/status:
    get:
      tags:
        - 系统状态
      summary: 获取防火墙状态
      responses:
        '200':
          description: 成功返回状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'

  /api/rules:
    get:
      tags:
        - 规则管理
      summary: 列出所有规则
      responses:
        '200':
          description: 规则列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleResponse'
    post:
      tags:
        - 规则管理
      summary: 创建新规则
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '200':
          description: 规则创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          description: 参数错误
        '500':
          description: 内部处理错误（如 GeoIP 下载失败或 eBPF 同步失败）

  /api/rules/{id}:
    delete:
      tags:
        - 规则管理
      summary: 删除指定规则
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '404':
          description: 规则不存在

components:
  schemas:
    Direction:
      type: string
      enum: [in, out]
      description: 流量方向

    Protocol:
      type: string
      enum: [tcp, udp, all]
      description: 协议类型

    Action:
      type: string
      enum: [block, pass]
      description: 匹配动作

    IpConfig:
      type: object
      oneOf:
        - $ref: '#/components/schemas/IpConfigAny'
        - $ref: '#/components/schemas/IpConfigCidr'
        - $ref: '#/components/schemas/IpConfigGeoip'
      discriminator:
        propertyName: ip_type

    IpConfigAny:
      type: object
      required: [ip_type]
      properties:
        ip_type:
          type: string
          enum: [any]

    IpConfigCidr:
      type: object
      required: [ip_type, ip]
      properties:
        ip_type:
          type: string
          enum: [cidr]
        ip:
          type: string
          description: CIDR 格式的 IP 段
          example: "1.2.3.4/32"

    IpConfigGeoip:
      type: object
      required: [ip_type, countries]
      properties:
        ip_type:
          type: string
          enum: [geoip]
        countries:
          type: array
          items:
            type: string
          description: ISO 国家代码列表
          example: ["CN", "US"]

    CreateRuleRequest:
      type: object
      required:
        - priority
        - direction
        - protocol
        - port_start
        - ip_type
        - action
      properties:
        priority:
          type: integer
          description: 优先级，值越大越优先匹配
        enabled:
          type: boolean
          default: true
        direction:
          $ref: '#/components/schemas/Direction'
        protocol:
          $ref: '#/components/schemas/Protocol'
        port_start:
          type: integer
          minimum: 0
          maximum: 65535
        port_end:
          type: integer
          minimum: 0
          maximum: 65535
          description: 结束端口，默认等于 port_start
        action:
          $ref: '#/components/schemas/Action'
      # ip_config fields are flattened via oneOf in actual implementation
      allOf:
        - $ref: '#/components/schemas/IpConfig'

    RuleResponse:
      type: object
      properties:
        id:
          type: integer
        priority:
          type: integer
        enabled:
          type: boolean
        direction:
          type: string
        protocol:
          type: string
        port_start:
          type: integer
        port_end:
          type: integer
        action:
          type: string
      allOf:
        - $ref: '#/components/schemas/IpConfig'

    StatusResponse:
      type: object
      properties:
        iface:
          type: string
        api_version:
          type: string
        rule_count:
          type: integer